+++ docs/core/extensions/snippets/logging/log-sampling/appsettings.json
@@ -4,7 +4,6 @@
       "Default": "Debug"
     }
   },
-
   "RandomProbabilisticSampler": {
     "Rules": [
       {
