+++ docs/core/extensions/snippets/logging/log-sampling/appsettings.noop.json
@@ -1,0 +1,11 @@
+{
+  "Logging": {
+    "RandomProbabilisticSampler": {
+      "Rules": [
+        {
+          "Probability": 1
+        }
+      ]
+    }
+  }
+}
