+++ docs/core/extensions/snippets/logging/log-sampling/appsettings.updated.json
@@ -1,0 +1,12 @@
+{
+  "Logging": {
+    "RandomProbabilisticSampler": {
+      "Rules": [
+        {
+          "Probability": 0.01,
+          "LogLevel": "Information"
+        }
+      ]
+    }
+  }
+}
