+++ docs/core/extensions/snippets/logging/log-sampling/file-config/appsettings.json
