+++ docs/core/extensions/snippets/logging/log-sampling/trace-based/Log.cs
@@ -1,6 +1,3 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
 using Microsoft.Extensions.Logging;
 
 namespace LogSamplingTraceBased;
