--- docs/core/extensions/snippets/logging/logging-sampling/appsettings.noop.json
@@ -1,11 +1,0 @@
-{
-    "Logging": {
-        "RandomProbabilisticSampler": {
-            "Rules": [
-                {
-                  "Probability": 1
-                }
-            ]
-        }
-    }
-}
