--- docs/core/extensions/snippets/logging/logging-sampling/appsettings.updated.json
@@ -1,12 +1,0 @@
-{
-    "Logging": {
-        "RandomProbabilisticSampler": {
-            "Rules": [
-                {
-                  "Probability": 0.01,
-                  "LogLevel": "Information"
-                }
-            ]
-        }
-    }
-}
