--- docs/core/extensions/snippets/logging/logging-sampling/code-config/appsettings.json
@@ -1,7 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Debug"
-    }
-  }
-}
