--- docs/core/extensions/snippets/logging/logging-sampling/trace-based/appsettings.json
@@ -1,7 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Trace"
-    }
-  }
-}
