+++ docs/fundamentals/toc.yml
@@ -1096,6 +1096,8 @@ items:
           - name: High-performance logging
             href: ../core/extensions/high-performance-logging.md
             displayName: high-performance logging,high-performance log,high-performance logging provider,high-performance log provider
+          - name: Log Sampling
+            href: ../../core/extensions/log-sampling.md
           - name: Console log formatting
             href: ../core/extensions/console-log-formatter.md
             displayName: console log formatting,console log formatter,console log formatting provider,console log formatter provider
