+++ docs/navigate/tools-diagnostics/toc.yml
@@ -363,8 +363,8 @@ items:
         href: ../../core/diagnostics/logging-tracing.md
       - name: ILogger Logging
         href: ../../core/extensions/logging.md
-      - name: Logging Sampling
-        href: ../../core/extensions/logging-sampling.md
+      - name: Log Sampling
+        href: ../../core/extensions/log-sampling.md
       - name: Observability with OpenTelemetry
         items:
           - name: Overview
